The Locker Room of Deep Dark Fantasy ♡

+++ src/dreamschool01/cscourse/chapter06/Student.java
@@ -19,7 +19,7 @@ public class Student {
 		System.out.println("국어 점수  :" + korean);
 		System.out.println("영어 점수 : " + english);
 		System.out.println("수학 점수 : " + math);
-		System.out.println("평균 점수 : " + average());
+		System.out.println("평균 점수 : " + (int)average());
 	}
 	void chageSutdentID(String newId) {
 		studentNum = newId;
